LyPy/requirements.txt
PyQt5>=5.15.0
requests>=2.31.0
winrt-runtime>=3.0.0
winrt-Windows.Media.Control>=3.0.0
winrt-Windows.Foundation>=3.0.0
winrt-Windows.Storage.Streams>=3.0.0
Pillow>=10.0.0
